chore(33212): fix ts configuration for Cypress
