+++ hivemq-edge-frontend/cypress/support/commands.ts
@@ -1,7 +1,28 @@
-import { getByTestId } from './commands/getByTestId.ts'
-import { getByAriaLabel } from './commands/getByAriaLabel.ts'
-import { checkAccessibility } from './commands/checkAccessibility.ts'
-import { clearInterceptList } from './commands/clearInterceptList.ts'
+/// <reference types="cypress" />
+import type { ContextObject } from 'axe-core'
+import type { Options } from 'cypress-axe'
+import { getByTestId } from './commands/getByTestId'
+import { getByAriaLabel } from './commands/getByAriaLabel'
+import { checkAccessibility } from './commands/checkAccessibility'
+import { clearInterceptList } from './commands/clearInterceptList'
+
+declare global {
+  namespace Cypress {
+    interface Chainable {
+      getByTestId(value: string): Chainable<JQuery<HTMLElement>>
+
+      getByAriaLabel(value: string): Chainable<JQuery<HTMLElement>>
+
+      checkAccessibility(
+        context?: string | Node | ContextObject | undefined,
+        options?: Options | undefined,
+        skipFailures?: boolean
+      ): Chainable<JQuery<HTMLElement>>
+
+      clearInterceptList(interceptAlias: string): Chainable<JQuery<HTMLElement>>
+    }
+  }
+}
 
 Cypress.Commands.add('getByTestId', getByTestId)
 Cypress.Commands.add('getByAriaLabel', getByAriaLabel)
