+++ hivemq-edge-frontend/cypress/support/commands/clearInterceptList.ts
@@ -1,5 +1,5 @@
 // @ts-ignore an import is not working
-import { Interception } from 'cypress/types/net-stubbing'
+import type { Interception } from 'cypress/types/net-stubbing'
 
 export const clearInterceptList = (interceptAlias: string): void => {
   cy.get(interceptAlias + '.all').then((browserRequests: Interception) => {
