+++ hivemq-edge-frontend/cypress/support/e2e.ts
@@ -10,6 +10,3 @@ import './commands'
 import installLogsCollector from 'cypress-terminal-report/src/installLogsCollector'
 
 installLogsCollector()
-
-// Alternatively you can use CommonJS syntax:
-// require('./commands')
