+++ hivemq-edge-frontend/src/modules/UnifiedNamespace/components/NamespaceDisplay.spec.cy.tsx
@@ -1,5 +1,3 @@
-// it works with the direct import
-// import 'cypress/cypress.d.ts'
 import NamespaceDisplay from './NamespaceDisplay.tsx'
 import { MOCK_BREADCRUMB, MOCK_NAMESPACE } from '@/__test-utils__/mocks.ts'
 
@@ -17,7 +15,6 @@ describe('NamespaceDisplay', () => {
     },
   ]
 
-  // it.each should be working now
   it.each(selectors)(
     (selector) => `should render properly ${selector.test}`,
     (selector) => {
