+++ hivemq-edge-frontend/tsconfig.app.json
@@ -24,11 +24,16 @@
     "noUncheckedSideEffectImports": true,
 
     // Alias for the imports; also in vite.config.ts
+    "types": ["cypress", "cypress-axe", "@percy/cypress", "cypress-real-events", "vite/client", "cypress-each"],
     "baseUrl": "./",
     "paths": {
       "@/*": ["src/*"],
       "@datahub/*": ["src/extensions/datahub/*"]
     }
   },
-  "include": ["src"]
+  "include": [
+    "src",
+    "node_modules/cypress/types/mocha/index.d.ts",
+    "node_modules/cypress/types/cypress-global-vars.d.ts"
+  ]
 }
