+++ hivemq-edge-frontend/tsconfig.json
@@ -1,4 +1,21 @@
 {
+  // All these need to be replicated from tsconfig.app.json, for Cypress to work
+  "compilerOptions": {
+    "moduleResolution": "bundler",
+    "module": "ESNext",
+    "jsx": "react-jsx",
+    "allowImportingTsExtensions": true,
+    "noEmit": true,
+
+    "types": ["cypress", "cypress-axe", "@percy/cypress", "cypress-real-events", "vite/client", "cypress-each"],
+    // Alias for the imports; also in vite.config.ts
+    "baseUrl": "./",
+    "paths": {
+      "@/*": ["src/*"],
+      "@datahub/*": ["src/extensions/datahub/*"]
+    }
+  },
   "files": [],
+
   "references": [{ "path": "./tsconfig.app.json" }, { "path": "./tsconfig.node.json" }]
 }
